KBC/MRT/BGKIncompressible parallelized


# Conflicts:
#	src/analysis/timeintegration-analysis/timeintegration-analysis-solver.cpp
#	src/library/natrium/solver/CFDSolver.cpp
#	src/library/natrium/solver/SolverConfiguration.h
#	src/test/collision/BGKStandard_test.cpp


Former-commit-id: 4ea1055ef318aee885f0b8c21964ee048a7e3808 [formerly 197a3edbc791d4de53cfcf2e053eb59294e19241]
Former-commit-id: e98ac3c29daf41b3a8a02478b4e3cf71eb86a984

+++ src/test/solver/ErrorStats_test.cpp
@@ -23,13 +23,13 @@ namespace natrium {
 BOOST_AUTO_TEST_SUITE(ErrorStats_test)
 
 BOOST_AUTO_TEST_CASE(ErrorStats_ConstructionAndFunctionality_test) {
-	cout << "ErrorStats_ConstructionAndFunctionality_test..." << endl;
+	pout << "ErrorStats_ConstructionAndFunctionality_test..." << endl;
 
 	// make solver object
 	const boost::filesystem::path natriumTmpDir("/tmp/natrium_test");
 	const boost::filesystem::path natriumTmpFile = natriumTmpDir
 			/ "test_errorstable.txt";
-	shared_ptr<SolverConfiguration> testConfiguration = make_shared<
+	boost::shared_ptr<SolverConfiguration> testConfiguration = boost::make_shared<
 			SolverConfiguration>();
 	testConfiguration->setOutputDirectory(natriumTmpDir.c_str());
 	testConfiguration->setCommandLineVerbosity(SILENT);
@@ -37,15 +37,17 @@ BOOST_AUTO_TEST_CASE(ErrorStats_ConstructionAndFunctionality_test) {
 	testConfiguration->setNumberOfTimeSteps(10);
 	size_t refinementLevel = 3;
 	double viscosity = 0.9;
-	shared_ptr<Benchmark<2> > testFlow = make_shared<
-			TaylorGreenVortex2D>(viscosity, refinementLevel);
+	boost::shared_ptr<Benchmark<2> > testFlow = boost::make_shared<TaylorGreenVortex2D>(
+			viscosity, refinementLevel);
 	BenchmarkCFDSolver<2> solver(testConfiguration, testFlow);
 
 	// make error stats object
-	if (boost::filesystem::is_directory(natriumTmpDir)) {
-		boost::filesystem::remove_all(natriumTmpDir);
+	if (is_MPI_rank_0()) {
+		if (boost::filesystem::is_directory(natriumTmpDir)) {
+			boost::filesystem::remove_all(natriumTmpDir);
+		}
+		boost::filesystem::create_directory(natriumTmpDir);
 	}
-	boost::filesystem::create_directory(natriumTmpDir);
 	ErrorStats<2> stats(&solver, natriumTmpFile.c_str());
 	BOOST_CHECK(boost::filesystem::exists(natriumTmpFile));
 
@@ -56,6 +58,7 @@ BOOST_AUTO_TEST_CASE(ErrorStats_ConstructionAndFunctionality_test) {
 
 	// count lines
 	std::ifstream file(natriumTmpFile.c_str());
+	MPI_sync();
 	size_t linecount = std::count(std::istreambuf_iterator<char>(file),
 			std::istreambuf_iterator<char>(), '\n');
 	BOOST_CHECK(linecount == 1);
@@ -64,20 +67,21 @@ BOOST_AUTO_TEST_CASE(ErrorStats_ConstructionAndFunctionality_test) {
 	stats.printNewLine();
 	solver.run();
 	stats.printNewLine();
+	MPI_sync();
 	BOOST_CHECK_EQUAL(stats.getIterationNumber(),
 			testConfiguration->getNumberOfTimeSteps());
 	linecount = std::count(std::istreambuf_iterator<char>(file),
 			std::istreambuf_iterator<char>(), '\n');
 	BOOST_CHECK(linecount > 1);
 
-	cout << "done" << endl;
+	pout << "done" << endl;
 }
 
 BOOST_AUTO_TEST_CASE(ErrorStats_FunctionsInSolverContext_test) {
-	cout << "ErrorStats_FunctionsInSolverContext_test..." << endl;
+	pout << "ErrorStats_FunctionsInSolverContext_test..." << endl;
 	// make solver object
 	const boost::filesystem::path natriumTmpDir("/tmp/natrium_test");
-	shared_ptr<SolverConfiguration> testConfiguration = make_shared<
+	boost::shared_ptr<SolverConfiguration> testConfiguration = boost::make_shared<
 			SolverConfiguration>();
 	testConfiguration->setOutputDirectory(natriumTmpDir.c_str());
 	testConfiguration->setCommandLineVerbosity(SILENT);
@@ -86,8 +90,8 @@ BOOST_AUTO_TEST_CASE(ErrorStats_FunctionsInSolverContext_test) {
 	testConfiguration->setOutputTableInterval(1);
 	size_t refinementLevel = 3;
 	double viscosity = 0.9;
-	shared_ptr<Benchmark<2> > testFlow = make_shared<
-			TaylorGreenVortex2D>(viscosity, refinementLevel);
+	boost::shared_ptr<Benchmark<2> > testFlow = boost::make_shared<TaylorGreenVortex2D>(
+			viscosity, refinementLevel);
 	BenchmarkCFDSolver<2> solver(testConfiguration, testFlow);
 
 	solver.run();
@@ -97,11 +101,12 @@ BOOST_AUTO_TEST_CASE(ErrorStats_FunctionsInSolverContext_test) {
 
 	// count lines
 	std::ifstream file(solver.getErrorStats()->getFilename());
+	MPI_sync();
 	size_t linecount = std::count(std::istreambuf_iterator<char>(file),
 			std::istreambuf_iterator<char>(), '\n');
 	BOOST_CHECK(linecount > 8);
 
-	cout << "done" << endl;
+	pout << "done" << endl;
 }
 
 BOOST_AUTO_TEST_SUITE_END()
