+++ src/test/solver/SolverStats_test.cpp
@@ -22,13 +22,13 @@ namespace natrium {
 BOOST_AUTO_TEST_SUITE(SolverStats_test)
 
 BOOST_AUTO_TEST_CASE(SolverStats_ConstructionAndFunctionality_test) {
-	cout << "SolverStats_ConstructionAndFunctionality_test..." << endl;
+	pout << "SolverStats_ConstructionAndFunctionality_test..." << endl;
 
 	// make solver object
 	const boost::filesystem::path natriumTmpDir("/tmp/natrium_test");
 	const boost::filesystem::path natriumTmpFile = natriumTmpDir
 			/ "testtable.txt";
-	shared_ptr<SolverConfiguration> testConfiguration = make_shared<
+	boost::shared_ptr<SolverConfiguration> testConfiguration = boost::make_shared<
 			SolverConfiguration>();
 	testConfiguration->setOutputDirectory(natriumTmpDir.c_str());
 	testConfiguration->setCommandLineVerbosity(SILENT);
@@ -36,15 +36,17 @@ BOOST_AUTO_TEST_CASE(SolverStats_ConstructionAndFunctionality_test) {
 	testConfiguration->setNumberOfTimeSteps(10);
 	size_t refinementLevel = 3;
 	double viscosity = 0.9;
-	shared_ptr<ProblemDescription<2> > testFlow = make_shared<
+	boost::shared_ptr<ProblemDescription<2> > testFlow = boost::make_shared<
 			SteadyPeriodicTestFlow2D>(viscosity, refinementLevel);
 	CFDSolver<2> solver(testConfiguration, testFlow);
 
 	// make solver stats object
-	if (boost::filesystem::is_directory(natriumTmpDir)) {
-		boost::filesystem::remove_all(natriumTmpDir);
+	if (is_MPI_rank_0()) {
+		if (boost::filesystem::is_directory(natriumTmpDir)) {
+			boost::filesystem::remove_all(natriumTmpDir);
+		}
+		boost::filesystem::create_directory(natriumTmpDir);
 	}
-	boost::filesystem::create_directory(natriumTmpDir);
 	SolverStats<2> stats(&solver, natriumTmpFile.c_str());
 	BOOST_CHECK(boost::filesystem::exists(natriumTmpFile));
 
@@ -54,6 +56,7 @@ BOOST_AUTO_TEST_CASE(SolverStats_ConstructionAndFunctionality_test) {
 	BOOST_CHECK(stats.isUpToDate());
 
 	// count lines
+	MPI_sync();
 	std::ifstream file(natriumTmpFile.c_str());
 	size_t linecount = std::count(std::istreambuf_iterator<char>(file),
 			std::istreambuf_iterator<char>(), '\n');
@@ -63,20 +66,21 @@ BOOST_AUTO_TEST_CASE(SolverStats_ConstructionAndFunctionality_test) {
 	stats.printNewLine();
 	solver.run();
 	stats.printNewLine();
+	MPI_sync();
 	BOOST_CHECK_EQUAL(stats.getIterationNumber(),
 			testConfiguration->getNumberOfTimeSteps());
 	linecount = std::count(std::istreambuf_iterator<char>(file),
 			std::istreambuf_iterator<char>(), '\n');
 	BOOST_CHECK(linecount > 1);
 
-	cout << "done" << endl;
+	pout << "done" << endl;
 }
 
 BOOST_AUTO_TEST_CASE(SolverStats_FunctionsInSolverContext_test) {
-	cout << "SolverStats_FunctionsInSolverContext_test..." << endl;
+	pout << "SolverStats_FunctionsInSolverContext_test..." << endl;
 	// make solver object
 	const boost::filesystem::path natriumTmpDir("/tmp/natrium_test");
-	shared_ptr<SolverConfiguration> testConfiguration = make_shared<
+	boost::shared_ptr<SolverConfiguration> testConfiguration = boost::make_shared<
 			SolverConfiguration>();
 	testConfiguration->setOutputDirectory(natriumTmpDir.c_str());
 	testConfiguration->setCommandLineVerbosity(SILENT);
@@ -85,7 +89,7 @@ BOOST_AUTO_TEST_CASE(SolverStats_FunctionsInSolverContext_test) {
 	testConfiguration->setOutputTableInterval(1);
 	size_t refinementLevel = 3;
 	double viscosity = 0.9;
-	shared_ptr<ProblemDescription<2> > testFlow = make_shared<
+	boost::shared_ptr<ProblemDescription<2> > testFlow = boost::make_shared<
 			SteadyPeriodicTestFlow2D>(viscosity, refinementLevel);
 	CFDSolver<2> solver(testConfiguration, testFlow);
 
@@ -100,7 +104,7 @@ BOOST_AUTO_TEST_CASE(SolverStats_FunctionsInSolverContext_test) {
 			std::istreambuf_iterator<char>(), '\n');
 	BOOST_CHECK(linecount > 8);
 
-	cout << "done" << endl;
+	pout << "done" << endl;
 }
 
 BOOST_AUTO_TEST_SUITE_END()
